+++ packages/client/src/services/logger.service.ts
@@ -1,4 +1,6 @@
-export const Logger = (name: string, enabled: boolean = false) => {
+export type LoggerLike = Pick<Console, 'debug' | 'error' | 'info' | 'log' | 'trace' | 'warn'>;
+
+export const Logger = (name: string, enabled: boolean = false): LoggerLike => {
   if (!enabled) {
     return {
       debug: () => {},
@@ -10,31 +12,15 @@ export const Logger = (name: string, enabled: boolean = false) => {
     };
   }
 
-  if (typeof window !== 'undefined') {
-    // Browser logger
-    return console;
-  }
-
-  const PinoLogger = require('pino');
-  const Pretty = require('pino-pretty');
-
-  return PinoLogger(
-    Pretty({
-      colorize: true,
-      translateTime: true,
-      ignore: 'pid,hostname',
-    }),
-  );
+  return console;
 };
 
-export type Logger = typeof console;
-
-export const handleError = (logger: Logger, message?: string) => (error: Error) => {
+export const handleError = (logger: LoggerLike, message?: string) => (error: Error) => {
   logger.error(error.message);
   throw new Error(`Error Occurred: ${message ?? error.message}`);
 };
 
-export const handleDebug = (logger: Logger, message?: string) => (res: any) => {
+export const handleDebug = (logger: LoggerLike, message?: string) => (res: any) => {
   logger.debug(message, res);
   return res;
 };