+++ playground/src/components/select-tenant/UseSelectTenant.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 
 export const useSelectTenant = () => {
   const [tenant, setTenant] = useState<Tenant>('davinci');
-  const [deployment, setDeployment] = useState<Deployment>('development');
+  const [deployment, setDeployment] = useState<Deployment>('staging');
   const [logger, setLogger] = useState<boolean>(true);
 
   return {